avoid a bunch of warnings about signedness/precision

+++ trunk/src/audio/datasource_filereader.h
@@ -12,7 +12,7 @@ class FileReaderDataSource: public IDataSource
 
 		void reset();
 		bool exhausted();
-		long getpos();
+		uint32 getpos();
 		uint32 getData(uint8* buffer, uint32 len);
 	private:
 		FILE* FileHandle;
